fix(pr-execute): address 5 findings from triage
- spec-conformance — PreCloseAction node ordering: moved after CheckReportNotClosed guard
- ci-status — Remove actor_id from debug log to resolve CodeQL HIGH
- pr-body-format — Closes #1854 moved to top of PR body
- test-coverage — Add test_custom_call_out_factory_is_used for custom bundle injection
- notes-currency — Move TestSvcCloseCaseUseCase to dedicated test_close_case_trigger.py
